requirement.txt
python-dotenv
streamlit
langchain_google_genai
langchain